// src/shared/api/v1/artifact.ts
export interface CreateArtifactRequest {
  /**
   * @generated from protobuf field: string workflow_run_backend_id = 1;
   */
  workflowRunBackendId: string
  /**
   * @generated from protobuf field: string workflow_job_run_backend_id = 2;
   */
  workflowJobRunBackendId: string
  /**
   * @generated from protobuf field: string name = 3;
   */
  name: string
  /**
   * @generated from protobuf field: google.protobuf.Timestamp expires_at = 4;
   */
  expiresAt?: Timestamp
  /**
   * @generated from protobuf field: int32 version = 5;
   */
  version: number
}
// ...
export interface FinalizeArtifactRequest {
  /**
   * @generated from protobuf field: string workflow_run_backend_id = 1;
   */
  workflowRunBackendId: string
  /**
   * @generated from protobuf field: string workflow_job_run_backend_id = 2;
   */
  workflowJobRunBackendId: string
  /**
   * @generated from protobuf field: string name = 3;
   */
  name: string
  /**
   * @generated from protobuf field: int64 size = 4;
   */
  size: string
  /**
   * @generated from protobuf field: google.protobuf.StringValue hash = 5;
   */
  hash?: StringValue // optional
}
// ...
class CreateArtifactRequest$Type extends MessageType<CreateArtifactRequest> {
  constructor() {
    super('github.actions.results.api.v1.CreateArtifactRequest', [
      {
        no: 1,
        name: 'workflow_run_backend_id',
        kind: 'scalar',
        T: 9 /*ScalarType.STRING*/
      },
      {
        no: 2,
        name: 'workflow_job_run_backend_id',
        kind: 'scalar',
        T: 9 /*ScalarType.STRING*/
      },
      {no: 3, name: 'name', kind: 'scalar', T: 9 /*ScalarType.STRING*/},
      {no: 4, name: 'expires_at', kind: 'message', T: () => Timestamp},
      {no: 5, name: 'version', kind: 'scalar', T: 5 /*ScalarType.INT32*/}
    ])
  }
  create(value?: PartialMessage<CreateArtifactRequest>): CreateArtifactRequest {
    const message = {
      workflowRunBackendId: '',
      workflowJobRunBackendId: '',
      name: '',
      version: 0
    }
    globalThis.Object.defineProperty(message, MESSAGE_TYPE, {
      enumerable: false,
      value: this
    })
    if (value !== undefined)
      reflectionMergePartial<CreateArtifactRequest>(this, message, value)
    return message
  }
  internalBinaryRead(
    reader: IBinaryReader,
    length: number,
    options: BinaryReadOptions,
    target?: CreateArtifactRequest
  ): CreateArtifactRequest {
    let message = target ?? this.create(),
      end = reader.pos + length
    while (reader.pos < end) {
      let [fieldNo, wireType] = reader.tag()
      switch (fieldNo) {
        case /* string workflow_run_backend_id */ 1:
          message.workflowRunBackendId = reader.string()
          break
        case /* string workflow_job_run_backend_id */ 2:
          message.workflowJobRunBackendId = reader.string()
          break
        case /* string name */ 3:
          message.name = reader.string()
          break
        case /* google.protobuf.Timestamp expires_at */ 4:
          message.expiresAt = Timestamp.internalBinaryRead(
            reader,
            reader.uint32(),
            options,
            message.expiresAt
          )
          break
        case /* int32 version */ 5:
          message.version = reader.int32()
          break
        default:
          let u = options.readUnknownField
          if (u === 'throw')
            throw new globalThis.Error(
              `Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`
            )
          let d = reader.skip(wireType)
          if (u !== false)
            (u === true ? UnknownFieldHandler.onRead : u)(
              this.typeName,
              message,
              fieldNo,
              wireType,
              d
            )
      }
    }
    return message
  }
  internalBinaryWrite(
    message: CreateArtifactRequest,
    writer: IBinaryWriter,
    options: BinaryWriteOptions
  ): IBinaryWriter {
    /* string workflow_run_backend_id = 1; */
    if (message.workflowRunBackendId !== '')
      writer
        .tag(1, WireType.LengthDelimited)
        .string(message.workflowRunBackendId)
    /* string workflow_job_run_backend_id = 2; */
    if (message.workflowJobRunBackendId !== '')
      writer
        .tag(2, WireType.LengthDelimited)
        .string(message.workflowJobRunBackendId)
    /* string name = 3; */
    if (message.name !== '')
      writer.tag(3, WireType.LengthDelimited).string(message.name)
    /* google.protobuf.Timestamp expires_at = 4; */
    if (message.expiresAt)
      Timestamp.internalBinaryWrite(
        message.expiresAt,
        writer.tag(4, WireType.LengthDelimited).fork(),
        options
      ).join()
    /* int32 version = 5; */
    if (message.version !== 0)
      writer.tag(5, WireType.Varint).int32(message.version)
    let u = options.writeUnknownFields
    if (u !== false)
      (u == true ? UnknownFieldHandler.onWrite : u)(
        this.typeName,
        message,
        writer
      )
    return writer
  }
}
/**
 * @generated MessageType for protobuf message github.actions.results.api.v1.CreateArtifactRequest
 */
export const CreateArtifactRequest = new CreateArtifactRequest$Type()
// ...
class FinalizeArtifactRequest$Type extends MessageType<FinalizeArtifactRequest> {
  constructor() {
    super('github.actions.results.api.v1.FinalizeArtifactRequest', [
      {
        no: 1,
        name: 'workflow_run_backend_id',
        kind: 'scalar',
        T: 9 /*ScalarType.STRING*/
      },
      {
        no: 2,
        name: 'workflow_job_run_backend_id',
        kind: 'scalar',
        T: 9 /*ScalarType.STRING*/
      },
      {no: 3, name: 'name', kind: 'scalar', T: 9 /*ScalarType.STRING*/},
      {no: 4, name: 'size', kind: 'scalar', T: 3 /*ScalarType.INT64*/},
      {no: 5, name: 'hash', kind: 'message', T: () => StringValue}
    ])
  }
  create(
    value?: PartialMessage<FinalizeArtifactRequest>
  ): FinalizeArtifactRequest {
    const message = {
      workflowRunBackendId: '',
      workflowJobRunBackendId: '',
      name: '',
      size: '0'
    }
    globalThis.Object.defineProperty(message, MESSAGE_TYPE, {
      enumerable: false,
      value: this
    })
    if (value !== undefined)
      reflectionMergePartial<FinalizeArtifactRequest>(this, message, value)
    return message
  }
  internalBinaryRead(
    reader: IBinaryReader,
    length: number,
    options: BinaryReadOptions,
    target?: FinalizeArtifactRequest
  ): FinalizeArtifactRequest {
    let message = target ?? this.create(),
      end = reader.pos + length
    while (reader.pos < end) {
      let [fieldNo, wireType] = reader.tag()
      switch (fieldNo) {
        case /* string workflow_run_backend_id */ 1:
          message.workflowRunBackendId = reader.string()
          break
        case /* string workflow_job_run_backend_id */ 2:
          message.workflowJobRunBackendId = reader.string()
          break
        case /* string name */ 3:
          message.name = reader.string()
          break
        case /* int64 size */ 4:
          message.size = reader.int64().toString()
          break
        case /* google.protobuf.StringValue hash */ 5:
          message.hash = StringValue.internalBinaryRead(
            reader,
            reader.uint32(),
            options,
            message.hash
          )
          break
        default:
          let u = options.readUnknownField
          if (u === 'throw')
            throw new globalThis.Error(
              `Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`
            )
          let d = reader.skip(wireType)
          if (u !== false)
            (u === true ? UnknownFieldHandler.onRead : u)(
              this.typeName,
              message,
              fieldNo,
              wireType,
              d
            )
      }
    }
    return message
  }
  internalBinaryWrite(
    message: FinalizeArtifactRequest,
    writer: IBinaryWriter,
    options: BinaryWriteOptions
  ): IBinaryWriter {
    /* string workflow_run_backend_id = 1; */
    if (message.workflowRunBackendId !== '')
      writer
        .tag(1, WireType.LengthDelimited)
        .string(message.workflowRunBackendId)
    /* string workflow_job_run_backend_id = 2; */
    if (message.workflowJobRunBackendId !== '')
      writer
        .tag(2, WireType.LengthDelimited)
        .string(message.workflowJobRunBackendId)
    /* string name = 3; */
    if (message.name !== '')
      writer.tag(3, WireType.LengthDelimited).string(message.name)
    /* int64 size = 4; */
    if (message.size !== '0') writer.tag(4, WireType.Varint).int64(message.size)
    /* google.protobuf.StringValue hash = 5; */
    if (message.hash)
      StringValue.internalBinaryWrite(
        message.hash,
        writer.tag(5, WireType.LengthDelimited).fork(),
        options
      ).join()
    let u = options.writeUnknownFields
    if (u !== false)
      (u == true ? UnknownFieldHandler.onWrite : u)(
        this.typeName,
        message,
        writer
      )
    return writer
  }
}
/**
 * @generated MessageType for protobuf message github.actions.results.api.v1.FinalizeArtifactRequest
 */
export const FinalizeArtifactRequest = new FinalizeArtifactRequest$Type()