// src/shared/google/protobuf/wrappers.ts
export interface StringValue {
  /**
   * The string value.
   *
   * @generated from protobuf field: string value = 1;
   */
  value: string
}
// ...
class StringValue$Type extends MessageType<StringValue> {
  constructor() {
    super('google.protobuf.StringValue', [
      {no: 1, name: 'value', kind: 'scalar', T: 9 /*ScalarType.STRING*/}
    ])
  }
  /**
   * Encode `StringValue` to JSON string.
   */
  internalJsonWrite(
    message: StringValue,
    options: JsonWriteOptions
  ): JsonValue {
    return message.value
  }
  /**
   * Decode `StringValue` from JSON string.
   */
  internalJsonRead(
    json: JsonValue,
    options: JsonReadOptions,
    target?: StringValue
  ): StringValue {
    if (!target) target = this.create()
    target.value = this.refJsonReader.scalar(
      json,
      9,
      undefined,
      'value'
    ) as string
    return target
  }
  create(value?: PartialMessage<StringValue>): StringValue {
    const message = {value: ''}
    globalThis.Object.defineProperty(message, MESSAGE_TYPE, {
      enumerable: false,
      value: this
    })
    if (value !== undefined)
      reflectionMergePartial<StringValue>(this, message, value)
    return message
  }
  internalBinaryRead(
    reader: IBinaryReader,
    length: number,
    options: BinaryReadOptions,
    target?: StringValue
  ): StringValue {
    let message = target ?? this.create(),
      end = reader.pos + length
    while (reader.pos < end) {
      let [fieldNo, wireType] = reader.tag()
      switch (fieldNo) {
        case /* string value */ 1:
          message.value = reader.string()
          break
        default:
          let u = options.readUnknownField
          if (u === 'throw')
            throw new globalThis.Error(
              `Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`
            )
          let d = reader.skip(wireType)
          if (u !== false)
            (u === true ? UnknownFieldHandler.onRead : u)(
              this.typeName,
              message,
              fieldNo,
              wireType,
              d
            )
      }
    }
    return message
  }
  internalBinaryWrite(
    message: StringValue,
    writer: IBinaryWriter,
    options: BinaryWriteOptions
  ): IBinaryWriter {
    /* string value = 1; */
    if (message.value !== '')
      writer.tag(1, WireType.LengthDelimited).string(message.value)
    let u = options.writeUnknownFields
    if (u !== false)
      (u == true ? UnknownFieldHandler.onWrite : u)(
        this.typeName,
        message,
        writer
      )
    return writer
  }
}
/**
 * @generated MessageType for protobuf message google.protobuf.StringValue
 */
export const StringValue = new StringValue$Type()